requirements.txt
streamlit
pdfplumber
faiss-cpu
numpy
sentence-transformers
groq
